requirements/base.txt
# Base
pytz==2021.3
python-slugify==5.0.2
Pillow==9.0.0
psycopg2==2.9.3 --no-binary psycopg2

# Django
django==3.2

# Environment
django-environ==0.8.1

# Django REST Framework
djangorestframework==3.13.1
django-filter==2.4.0
django-cors-headers==3.5.0
drf-writable-nested==0.6.2


# JWT
pyjwt==2.3.0

# Passwords security
argon2-cffi==21.3.0

# Static files
whitenoise==5.3.0

# Celery
tornado==6.1
redis==4.1.2
django-redis==5.2.0
celery==5.2.3
flower==1.0.0
